horizon_dashboard/test-requirements.txt
coverage
django-nose
mox
netaddr
nose
pep8
pylint
sphinx